Setup kustomize on kyma alpha create module
